refactor(Controller): 기존 Controller의 역할을 Race 클래스로 분할하는 리팩토링 진행 및 클래스명 변경
